Show user code traceback for exceptions when running local entrypoint
